Remove ejson
