+++ README.md
@@ -1,3 +1,4 @@
-# ddp
+# DDP in Go
 
-MeteorJS DDP library for Golang
+[Meteor](https://meteor.com) DDP library for Go. This library allows Go applications to connect to Meteor applications, subscribe to Meteor publications, read from a cached Collection (similar to minimongo), and call Meteor methods on the server.
+
