+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/gopackage
+
+go 1.17
+
+require (
+	github.com/apex/log v1.9.0
+	github.com/stretchr/testify v1.7.0
+	golang.org/x/net v0.0.0-20210813160813-60bc85c4be6d
+)
